requirements.txt
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-WTF==1.2.2
python-dotenv==1.2.2
email-validator==2.3.0
alembic==1.18.4
